refactor[#26] : 이미지 다운로드 로직 커스텀훅으로 빼내어 비즈니스 로직과 ui 로직 분리
