+++ app/src/features/snippet/lib/index.ts
@@ -1,3 +1,4 @@
 export * from "./useCopySnippetImage";
 export * from "./useSnippetContent";
 export * from "./useSnippetTextArea";
+export * from "./useSaveSnippetImage";
