+++ app/src/features/snippet/ui/styles.module.css
@@ -174,4 +174,5 @@
   z-index: 999999;
   position: absolute;
   background: inherit;
+  padding-top: 3rem;
 }
